Magic Quotes Fix
Instead of removing the magic quotes statement, I simply added a check to it. Some people still want to run very old code (like me) on very old php versions. If removed, this file will not run on anything php 5.3 and older. So by adding a check instead of removing, more people will be able to enjoy the CMS.